Address browser picker review feedback
